Moved components from homepage to access.
